# apps/registrant/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Conference Attendee Registration App
  version: 0.1.0
  description: >
    QR-based agenda check-in, attendee summary reporting, and invitation
    email dispatch for conference attendees.
servers:
  - url: /
security:
  - jwtAssertion: []
paths:
...
  /attendees/sync:
    post:
      summary: Sync attendee summary to the reporting sheet
      operationId: syncAttendees
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendeeSyncPayload"
      responses:
        "200":
          description: Synced successfully
        "400":
          description: Missing or invalid payload
        "500":
          description: Internal server error
components:
  securitySchemes:
    jwtAssertion:
      type: apiKey
      in: header
      name: x-jwt-assertion
      description: >
        Invoker identity JWT. The payload is decoded for the caller's email;
        the signature is not verified by this service (trust boundary is the
        gateway in front of it).
  schemas:
...
    AttendeeSyncPayload:
      type: object
      properties:
        timeZoneOffset:
          type: number
          description: Time offset in hours of the invoker
      required: [timeZoneOffset]